fix: PR #480 - Path normalization and file structure refactoring
High Priority:
- Fixed path normalization regex to handle paths without trailing slash
- Changed patterns from requiring trailing slash to working with or without
- Example: /users/123 now correctly normalizes to /users/{username}

File Structure Refactoring:
- Split 976-line exceptions.py into focused modules:
  - pii.py: PIIMasker class (6.4KB)
  - providers.py: Message/Suggestion/Link providers (12.7KB)
  - error_models.py: Data models (1.2KB)
  - exceptions.py: Only exception classes (570 lines)
- Follows single responsibility principle
- Improves maintainability

Test Coverage:
- Added tests for paths without trailing slash
- Added edge case tests for all pattern types
- 14 new test cases covering regex fix

Pattern Fix Example:
Before: (r'/users/([^/]+)/', '/users/{username}/')
After: (r'(/users/)[^/]+', r'\1{username}')

All imports updated, all tests passing ✅
