+++ apps/api/app/core/exceptions.py
@@ -25,6 +25,9 @@ from fastapi import HTTPException, status
 
 from ..core.logging import get_logger
 from ..middleware.correlation_middleware import get_correlation_id
+from .pii import PIIMasker
+from .error_models import ErrorSuggestion, RemediationLink, ErrorDetails
+from .providers import ErrorMessageProvider, SuggestionProvider, RemediationLinkProvider
 
 logger = get_logger(__name__)
 
@@ -170,25 +173,6 @@ ERROR_HTTP_MAPPING: Dict[ErrorCode, int] = {
 }
 
 
-class ErrorSuggestion(BaseModel):
-    """Actionable suggestion for error remediation."""
-    en: str = Field(description="English suggestion text")
-    tr: str = Field(description="Turkish suggestion text")
-
-
-class RemediationLink(BaseModel):
-    """Documentation or resource link for error remediation."""
-    title: str = Field(description="Link title")
-    url: str = Field(description="Link URL")
-
-
-class ErrorDetails(BaseModel):
-    """Detailed error information."""
-    component: Optional[str] = Field(default=None, description="Component where error occurred")
-    exception_class: Optional[str] = Field(default=None, description="Exception class name")
-    phase: Optional[str] = Field(default=None, description="Processing phase")
-    file_format: Optional[str] = Field(default=None, description="File format if relevant")
-    param: Optional[Dict[str, Any]] = Field(default=None, description="Sanitized parameters")
 
 
 class ErrorResponse(BaseModel):
@@ -204,159 +188,6 @@ class ErrorResponse(BaseModel):
     job_id: Optional[str] = Field(default=None, description="Job ID if applicable")
     timestamp: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
 
-
-class PIIMasker:
-    """Utility class for masking PII in logs and error messages."""
-    
-    @staticmethod
-    def _is_valid_tc_kimlik_no(number: str) -> bool:
-        """Validate Turkish TC Kimlik No with checksum algorithm.
-        
-        TC Kimlik No validation rules:
-        - Must be 11 digits
-        - First digit cannot be 0
-        - 10th digit = ((sum of odd positions * 7) - sum of even positions) mod 10
-        - 11th digit = sum of first 10 digits mod 10
-        """
-        if not number.isdigit() or len(number) != 11 or number[0] == '0':
-            return False
-        
-        digits = [int(d) for d in number]
-        
-        # Calculate 10th digit checksum
-        odd_sum = sum(digits[i] for i in range(0, 9, 2))  # positions 1,3,5,7,9
-        even_sum = sum(digits[i] for i in range(1, 8, 2))  # positions 2,4,6,8
-        tenth_digit = (odd_sum * 7 - even_sum) % 10
-        
-        if digits[9] != tenth_digit:
-            return False
-        
-        # Calculate 11th digit checksum
-        eleventh_digit = sum(digits[:10]) % 10
-        
-        return digits[10] == eleventh_digit
-    
-    @staticmethod
-    def _is_valid_credit_card(number: str) -> bool:
-        """Validate credit card number using Luhn algorithm.
-        
-        Luhn algorithm:
-        1. Double every second digit from right to left
-        2. If doubling results in a number > 9, sum the digits
-        3. Sum all digits
-        4. Valid if sum mod 10 == 0
-        """
-        # Remove spaces and hyphens
-        cleaned = re.sub(r'[\s-]', '', number)
-        
-        if not cleaned.isdigit() or len(cleaned) < 13 or len(cleaned) > 19:
-            return False
-        
-        def luhn_checksum(card_number: str) -> int:
-            digits = [int(d) for d in card_number]
-            # Reverse for easier processing
-            digits.reverse()
-            
-            total = 0
-            for i, digit in enumerate(digits):
-                if i % 2 == 1:  # Every second digit (from right)
-                    doubled = digit * 2
-                    if doubled > 9:
-                        # Sum the digits (e.g., 16 -> 1 + 6 = 7)
-                        total += doubled // 10 + doubled % 10
-                    else:
-                        total += doubled
-                else:
-                    total += digit
-            
-            return total % 10
-        
-        return luhn_checksum(cleaned) == 0
-    
-    @classmethod
-    def _mask_tc_kimlik_no(cls, match):
-        """Mask Turkish TC Kimlik No only if valid."""
-        number = match.group(0)
-        if cls._is_valid_tc_kimlik_no(number):
-            return '[tc_no redacted]'
-        return number
-    
-    @classmethod
-    def _mask_credit_card(cls, match):
-        """Mask credit card only if valid according to Luhn algorithm."""
-        number = match.group(0)
-        if cls._is_valid_credit_card(number):
-            return '[card redacted]'
-        return number
-    
-    # PII patterns for masking
-    PATTERNS = {
-        'email': (r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b', '[email redacted]'),
-        'phone': (r'\b(?:\+?90|0)?(?:\s*\(?5\d{2}\)?[\s.-]?\d{3}[\s.-]?\d{2}[\s.-]?\d{2}|\d{3}[\s.-]?\d{3}[\s.-]?\d{4})\b', '[phone redacted]'),
-        'jwt': (r'eyJ[A-Za-z0-9_-]+\.[A-Za-z0-9_-]+\.[A-Za-z0-9_-]+', '[token redacted]'),
-        'api_key': (r'\b(?:api[_-]?key|apikey|token)["\']?\s*[:=]\s*["\']?([A-Za-z0-9_-]{20,})["\']?', '[api_key redacted]'),
-        'ip_address': (r'\b(?:\d{1,3}\.){3}\d{1,3}\b', '[ip redacted]'),
-        'home_dir': (r'(?:/home/[^/\s]+|/Users/[^/\s]+|C:\\Users\\[^\\]+)', '[path redacted]'),
-    }
-    
-    # Special patterns that require validation
-    VALIDATION_PATTERNS = {
-        'credit_card': (r'\b(?:\d{4}[\s-]?){3,4}\d{1,4}\b', _mask_credit_card),
-        'tc_kimlik_no': (r'\b\d{11}\b', _mask_tc_kimlik_no),
-    }
-    
-    @classmethod
-    def mask_text(cls, text: str) -> str:
-        """Mask PII in text."""
-        if not text:
-            return text
-        
-        masked = text
-        
-        # Apply simple patterns first
-        for pattern_name, (pattern, replacement) in cls.PATTERNS.items():
-            masked = re.sub(pattern, replacement, masked, flags=re.IGNORECASE)
-        
-        # Apply validation-based patterns
-        for pattern_name, (pattern, validator_func) in cls.VALIDATION_PATTERNS.items():
-            # Pass the class method with cls bound
-            masked = re.sub(pattern, lambda m: validator_func.__func__(cls, m), masked)
-        
-        return masked
-    
-    @classmethod
-    def _mask_recursive(cls, obj: Any) -> Any:
-        """Helper to recursively mask any data structure."""
-        if isinstance(obj, str):
-            return cls.mask_text(obj)
-        elif isinstance(obj, dict):
-            return cls.mask_dict(obj)
-        elif isinstance(obj, list):
-            return [cls._mask_recursive(item) for item in obj]
-        elif isinstance(obj, tuple):
-            return tuple(cls._mask_recursive(item) for item in obj)
-        elif isinstance(obj, set):
-            return {cls._mask_recursive(item) for item in obj}
-        else:
-            # For other types (int, float, bool, None, etc.), return as-is
-            return obj
-    
-    @classmethod
-    def mask_dict(cls, data: Dict[str, Any]) -> Dict[str, Any]:
-        """Recursively mask PII in dictionary."""
-        if not data:
-            return data
-        
-        masked = {}
-        for key, value in data.items():
-            # Mask sensitive keys entirely
-            if any(sensitive in key.lower() for sensitive in ['password', 'secret', 'token', 'key', 'auth']):
-                masked[key] = '[redacted]'
-            else:
-                # Use recursive helper for all value types
-                masked[key] = cls._mask_recursive(value)
-        
-        return masked
 
 
 class FreeCADErrorPatternMatcher:
@@ -425,243 +256,6 @@ class FreeCADErrorPatternMatcher:
         
         return None
 
-
-class ErrorMessageProvider:
-    """Provider for bilingual error messages."""
-    
-    MESSAGES = {
-        ErrorCode.AI_AMBIGUOUS: {
-            "en": "The prompt is ambiguous and requires clarification.",
-            "tr": "İstem belirsiz ve açıklama gerektiriyor."
-        },
-        ErrorCode.AI_HINT_REQUIRED: {
-            "en": "Additional information is required to process this request.",
-            "tr": "Bu isteği işlemek için ek bilgi gerekiyor."
-        },
-        ErrorCode.FC_GEOM_INVALID_SHAPE: {
-            "en": "The model geometry is invalid (non-manifold or self-intersecting) and cannot be processed.",
-            "tr": "Model geometrisi geçersiz (manifold değil veya kendisiyle kesişiyor) ve işlenemiyor."
-        },
-        ErrorCode.FC_BOOLEAN_FAILED: {
-            "en": "Boolean operation failed due to invalid geometry or coplanar surfaces.",
-            "tr": "Boolean işlemi geçersiz geometri veya eş düzlemli yüzeyler nedeniyle başarısız oldu."
-        },
-        ErrorCode.FC_FILLET_CHAMFER_FAILED: {
-            "en": "Fillet/Chamfer operation failed; radius may exceed adjacent edge limits.",
-            "tr": "Kordon/Pah işlemi başarısız oldu; yarıçap komşu kenar sınırlarını aşmış olabilir."
-        },
-        ErrorCode.FC_SKETCH_OVERCONSTRAINED: {
-            "en": "Sketch is over-constrained; some constraints conflict.",
-            "tr": "Eskiz aşırı kısıtlanmış; bazı kısıtlar birbiriyle çakışıyor."
-        },
-        ErrorCode.FC_SKETCH_UNDERCONSTRAINED: {
-            "en": "Sketch is under-constrained and has degrees of freedom.",
-            "tr": "Eskiz yetersiz kısıtlanmış ve serbestlik dereceleri var."
-        },
-        ErrorCode.FC_IMPORT_STEP_FAILED: {
-            "en": "Failed to import the STEP file. The file may be corrupted or use unsupported entities.",
-            "tr": "STEP dosyası içe aktarılamadı. Dosya bozuk olabilir veya desteklenmeyen varlıklar içeriyor olabilir."
-        },
-        ErrorCode.FC_EXPORT_STL_FAILED: {
-            "en": "Failed to export STL file. The model may have invalid geometry.",
-            "tr": "STL dosyası dışa aktarılamadı. Model geçersiz geometriye sahip olabilir."
-        },
-        ErrorCode.FC_A4_UNSOLVED: {
-            "en": "Assembly constraints cannot be solved; check LCS alignment and cyclic dependencies.",
-            "tr": "Montaj kısıtları çözülemedi; LCS hizalamasını ve döngüsel bağımlılıkları kontrol edin."
-        },
-        ErrorCode.FC_A4_LINK_SCOPE: {
-            "en": "Assembly links are out of scope or LCS is missing.",
-            "tr": "Montaj bağlantıları kapsam dışında veya LCS eksik."
-        },
-        ErrorCode.FC_MESH_FAILED: {
-            "en": "Mesh generation failed due to invalid geometry or non-manifold edges.",
-            "tr": "Ağ oluşturma geçersiz geometri veya manifold olmayan kenarlar nedeniyle başarısız oldu."
-        },
-        ErrorCode.FC_RECOMPUTE_FAILED: {
-            "en": "Document recompute failed due to cyclic dependencies or invalid operations.",
-            "tr": "Belge yeniden hesaplama döngüsel bağımlılıklar veya geçersiz işlemler nedeniyle başarısız oldu."
-        },
-        ErrorCode.FC_TOPONAMING_UNSTABLE: {
-            "en": "Topological references were lost after model changes.",
-            "tr": "Model değişikliklerinden sonra topolojik referanslar kayboldu."
-        },
-        ErrorCode.TIMEOUT_WORKER: {
-            "en": "Processing timed out. The model is too complex or system is overloaded.",
-            "tr": "İşleme zaman aşımına uğradı. Model çok karmaşık veya sistem aşırı yük altında."
-        },
-        ErrorCode.STORAGE_QUOTA_EXCEEDED: {
-            "en": "Storage quota exceeded. Please free up space or upgrade your plan.",
-            "tr": "Depolama kotası aşıldı. Lütfen alan boşaltın veya planınızı yükseltin."
-        },
-        ErrorCode.VALIDATION_RANGE_VIOLATION: {
-            "en": "Value is outside the acceptable range.",
-            "tr": "Değer kabul edilebilir aralığın dışında."
-        },
-        ErrorCode.VALIDATION_CONFLICT: {
-            "en": "Conflicting parameters detected.",
-            "tr": "Çakışan parametreler tespit edildi."
-        },
-    }
-    
-    @classmethod
-    def get_message(cls, error_code: ErrorCode) -> Tuple[str, str]:
-        """Get bilingual error message for error code."""
-        messages = cls.MESSAGES.get(error_code, {})
-        return (
-            messages.get("en", f"Error: {error_code.value}"),
-            messages.get("tr", f"Hata: {error_code.value}")
-        )
-
-
-class SuggestionProvider:
-    """Provider for actionable error suggestions."""
-    
-    SUGGESTIONS = {
-        ErrorCode.FC_GEOM_INVALID_SHAPE: [
-            ErrorSuggestion(
-                en="Heal geometry and remove self-intersections; ensure solids are closed/manifold.",
-                tr="Geometriyi iyileştirin ve kendi kendini kesişmeleri kaldırın; katıların kapalı/manifold olduğundan emin olun."
-            ),
-            ErrorSuggestion(
-                en="Use 'Refine shape' after boolean operations.",
-                tr="Boolean işlemlerden sonra 'Refine shape' kullanın."
-            ),
-        ],
-        ErrorCode.FC_FILLET_CHAMFER_FAILED: [
-            ErrorSuggestion(
-                en="Reduce fillet radius below the smallest adjacent edge length.",
-                tr="Kordon yarıçapını en küçük komşu kenar uzunluğunun altına indirin."
-            ),
-            ErrorSuggestion(
-                en="Increase wall thickness to accommodate larger fillets.",
-                tr="Daha büyük kordonları barındırmak için duvar kalınlığını artırın."
-            ),
-        ],
-        ErrorCode.FC_SKETCH_OVERCONSTRAINED: [
-            ErrorSuggestion(
-                en="Remove redundant constraints and apply dimensional constraints incrementally.",
-                tr="Gereksiz kısıtları kaldırın ve boyutsal kısıtları kademeli uygulayın."
-            ),
-            ErrorSuggestion(
-                en="Check for conflicting dimensional and geometric constraints.",
-                tr="Çakışan boyutsal ve geometrik kısıtları kontrol edin."
-            ),
-        ],
-        ErrorCode.FC_IMPORT_STEP_FAILED: [
-            ErrorSuggestion(
-                en="Export STEP as AP214/AP242 format for better compatibility.",
-                tr="Daha iyi uyumluluk için STEP'i AP214/AP242 formatında dışa aktarın."
-            ),
-            ErrorSuggestion(
-                en="Ensure units are set to millimeters and run a CAD repair tool before upload.",
-                tr="Birimlerin milimetre olarak ayarlandığından emin olun ve yüklemeden önce CAD onarım aracı çalıştırın."
-            ),
-        ],
-        ErrorCode.FC_MESH_FAILED: [
-            ErrorSuggestion(
-                en="Enable mesh refinement and reduce feature complexity.",
-                tr="Ağ iyileştirmeyi etkinleştirin ve özellik karmaşıklığını azaltın."
-            ),
-            ErrorSuggestion(
-                en="Fix non-manifold edges and ensure watertight geometry.",
-                tr="Manifold olmayan kenarları düzeltin ve su geçirmez geometri sağlayın."
-            ),
-        ],
-        ErrorCode.FC_A4_UNSOLVED: [
-            ErrorSuggestion(
-                en="Ensure each part has an LCS (Local Coordinate System).",
-                tr="Her parçanın bir LCS'si (Yerel Koordinat Sistemi) olduğundan emin olun."
-            ),
-            ErrorSuggestion(
-                en="Solve constraints stepwise and avoid circular dependencies.",
-                tr="Kısıtları adım adım çözün ve döngüsel bağımlılıklardan kaçının."
-            ),
-        ],
-        ErrorCode.TIMEOUT_WORKER: [
-            ErrorSuggestion(
-                en="Simplify model by reducing feature count or fillet complexity.",
-                tr="Özellik sayısını veya kordon karmaşıklığını azaltarak modeli basitleştirin."
-            ),
-            ErrorSuggestion(
-                en="Split complex assemblies into smaller subassemblies.",
-                tr="Karmaşık montajları daha küçük alt montajlara bölün."
-            ),
-        ],
-        ErrorCode.VALIDATION_RANGE_VIOLATION: [
-            ErrorSuggestion(
-                en="Check minimum and maximum values for this parameter.",
-                tr="Bu parametre için minimum ve maksimum değerleri kontrol edin."
-            ),
-            ErrorSuggestion(
-                en="Ensure values meet manufacturing constraints (e.g., minimum wall thickness >= 1.5mm).",
-                tr="Değerlerin üretim kısıtlarını karşıladığından emin olun (örn. minimum duvar kalınlığı >= 1.5mm)."
-            ),
-        ],
-    }
-    
-    @classmethod
-    def get_suggestions(cls, error_code: ErrorCode) -> List[ErrorSuggestion]:
-        """Get suggestions for error code."""
-        return cls.SUGGESTIONS.get(error_code, [])
-
-
-class RemediationLinkProvider:
-    """Provider for remediation documentation links."""
-    
-    LINKS = {
-        ErrorCode.FC_GEOM_INVALID_SHAPE: [
-            RemediationLink(
-                title="FreeCAD Geometry Cleanup",
-                url="https://wiki.freecad.org/Part_RefineShape"
-            ),
-            RemediationLink(
-                title="BRep Validity and Healing",
-                url="https://wiki.freecad.org/Part_Workbench"
-            ),
-        ],
-        ErrorCode.FC_FILLET_CHAMFER_FAILED: [
-            RemediationLink(
-                title="Fillet Best Practices",
-                url="https://wiki.freecad.org/PartDesign_Fillet"
-            ),
-        ],
-        ErrorCode.FC_SKETCH_OVERCONSTRAINED: [
-            RemediationLink(
-                title="Sketcher Constraints Guide",
-                url="https://wiki.freecad.org/Sketcher_Workbench"
-            ),
-        ],
-        ErrorCode.FC_IMPORT_STEP_FAILED: [
-            RemediationLink(
-                title="STEP Import Tips",
-                url="https://wiki.freecad.org/Import_Export"
-            ),
-        ],
-        ErrorCode.FC_MESH_FAILED: [
-            RemediationLink(
-                title="Mesh Best Practices",
-                url="https://wiki.freecad.org/Mesh_Workbench"
-            ),
-        ],
-        ErrorCode.FC_A4_UNSOLVED: [
-            RemediationLink(
-                title="Assembly4 Documentation",
-                url="https://wiki.freecad.org/Assembly4_Workbench"
-            ),
-        ],
-        ErrorCode.TIMEOUT_WORKER: [
-            RemediationLink(
-                title="Performance Optimization Tips",
-                url="https://wiki.freecad.org/Performance_tips"
-            ),
-        ],
-    }
-    
-    @classmethod
-    def get_links(cls, error_code: ErrorCode) -> List[RemediationLink]:
-        """Get remediation links for error code."""
-        return cls.LINKS.get(error_code, [])
 
 
 class EnterpriseException(Exception):
